Add files via upload
This is a project that secures the content of a file by encrypting the data multiple times and generating multiples keys. The same keys are then required to decrypt the file data. This multiple keys concept makes the encrypted file content very secure. A key concept that I have used is that the coding standards are all CERT recommendations and is therefore is very secure which increases the file security.